refactor  location, user, event

+++ src/main/java/com/event/hub/service/LocationService.java
@@ -3,7 +3,7 @@ package com.event.hub.service;
 import com.event.hub.db.LocationRepository;
 import com.event.hub.db.entity.LocationEntity;
 import com.event.hub.filter.LocationSearchFilter;
-import com.event.hub.model.location.Location;
+import com.event.hub.model.location.LocationDto;
 import com.event.hub.model.location.LocationMapper;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -13,60 +13,63 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
-@Service
+@Service("locationService")
 @RequiredArgsConstructor
 public class LocationService {
     private final LocationMapper locationMapper;
     private final LocationRepository locationRepository;
 
 
-    @Transactional
-    public Location createLocation(Location location) {
-        LocationEntity entity = locationMapper.toEntity(location);
+    public LocationDto createLocation(LocationDto locationDto) {
+        LocationEntity entity = locationMapper.toEntity(locationDto);
 
         return saveAndMappedToDomain(entity);
     }
 
-    @Transactional
-    public Location updateLocation(Long id, Location location) {
+    public LocationDto updateLocation(Long id, LocationDto locationDto) {
         LocationEntity existedLocation = findLocationById(id);
-        LocationEntity entity = locationMapper.toEntity(location);
+        LocationEntity entity = locationMapper.toEntity(locationDto);
         entity.setId(id);
         capacityNotLessThanBefore(entity.getCapacity(), existedLocation);
         return saveAndMappedToDomain(entity);
     }
 
-    @Transactional
-    public Location patchLocation(Long id, Location location) {
-        isSameIds(id, location.id());
+    public LocationDto patchLocation(Long id, LocationDto locationDto) {
+        isSameIds(id, locationDto.id());
         LocationEntity entity = findLocationById(id);
-        if (isUniqueNameAndAddressLocation(location)) {
-            Optional.ofNullable(location.name()).ifPresent(entity::setName);
-            Optional.ofNullable(location.address()).ifPresent(entity::setAddress);
+        if (isUniqueNameAndAddressLocation(locationDto)) {
+            Optional.ofNullable(locationDto.name()).ifPresent(entity::setName);
+            Optional.ofNullable(locationDto.address()).ifPresent(entity::setAddress);
         }
-        Optional.ofNullable(location.capacity()).ifPresent(capacity -> {
+        Optional.ofNullable(locationDto.capacity()).ifPresent(capacity -> {
             capacityNotLessThanBefore(capacity, entity);
             entity.setCapacity(capacity);
         });
-        Optional.ofNullable(location.description()).ifPresent(entity::setDescription);
+        Optional.ofNullable(locationDto.description()).ifPresent(entity::setDescription);
 
         return saveAndMappedToDomain(entity);
     }
 
     @Transactional
+    public LocationDto saveAndMappedToDomain(LocationEntity entity) {
+        return locationMapper.toDomain(locationRepository.save(entity));
+    }
+
+    @Transactional
     public void delete(Long id) {
-        existsLocationById(id);
-        locationRepository.deleteById(id);
+        if (existsLocationById(id)) {
+            locationRepository.deleteById(id);
+        }
     }
 
     @Transactional(readOnly = true)
-    public Location getLocationById(Long id) {
+    public LocationDto getLocationById(Long id) {
         LocationEntity entity = findLocationById(id);
         return locationMapper.toDomain(entity);
     }
 
     @Transactional(readOnly = true)
-    public Page<Location> getAllLocation(LocationSearchFilter filter) {
+    public Page<LocationDto> getAllLocation(LocationSearchFilter filter) {
         Page<LocationEntity> all = locationRepository.findAll(
                 filter.toSpecification(),
                 filter.toPageable()
@@ -74,39 +77,45 @@ public class LocationService {
         return all.map(locationMapper::toDomain);
     }
 
-    public boolean isUniqueNameAndAddressLocation(Location location) {
+    public boolean isUniqueNameAndAddressLocation(LocationDto locationDto) {
         return !locationRepository.existsByNameOrAddress(
-                location.name(),
-                location.address()
+                locationDto.name(),
+                locationDto.address()
         );
     }
 
-    private static void capacityNotLessThanBefore(Integer capacity, LocationEntity entity) {
-        if (entity.getCapacity() > capacity) {
-            throw new IllegalStateException("New Capacity can't less than old one");
+    public boolean validateLocationFromRequestedPlaces(Long locationId, Integer requestedPlace) {
+        if (locationRepository.existsByIdAndCapacityLessThan(locationId, requestedPlace)) {
+            throw new IllegalStateException("There are not enough places in the selected location ID=%s"
+                    .formatted(locationId)
+            );
         }
+        return true;
     }
 
-    private LocationEntity findLocationById(Long id) {
+    public boolean existsLocationById(Long id) {
+        if (!locationRepository.existsById(id)) {
+            throw new EntityNotFoundException("No such Location by ID:" + id);
+        }
+        return true;
+    }
+
+    public LocationEntity findLocationById(Long id) {
         return locationRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("No such Location by ID:" + id)
         );
     }
 
-    private Location saveAndMappedToDomain(LocationEntity entity) {
-        return locationMapper.toDomain(locationRepository.save(entity));
+    private void capacityNotLessThanBefore(Integer capacity, LocationEntity entity) {
+        if (entity.getCapacity() > capacity) {
+            throw new IllegalStateException("New Capacity can't less than old one");
+        }
     }
 
     private static void isSameIds(Long id, Long otherId) {
         if (!otherId.equals(id)) {
             throw new IllegalArgumentException("Location ID:%s and ID:%s must be same."
                     .formatted(otherId, id));
-        }
-    }
-
-    private void existsLocationById(Long id) {
-        if (!locationRepository.existsById(id)) {
-            throw new EntityNotFoundException("No such Location by ID:" + id);
         }
     }
 }
